SQLCatalog_deferFullTextIndex{,Activity}: Do not use getObject .
getObject is generally a shortcut for "return self", and on brains for
traversing to corresponding document.
Except... Not everywhere. portal_catalog for example defines a completely
unrelated method which takes a UID as argument and does a query retrieving
one document. Which is basically totally redundant with its __call__
method. Still, of course, it is used in several places.
Also, ERP5Form.Report defines yet another getObject method, taking an
acquisition context as argument and traversing to some other path.

So do not use getObject (at least until the code recovers some sanity and
consistency), but getPath + custom activity API. Oh well.