ERP5Type.Base: Make _getReindexAndActivateParameterDict honor getDefaultActivateParameterDict
It is set in constructInstance from received activate_kw, but is then if
ActiveObject.activate is not used - as is the case when reindexObject is
made an alias of recursiveReindexObject (ex: ERP5.Document.Delivery), as
the latter calls recurseCallMethod which uses ActivityTool.activateObject
which (purposefully, visibly) ignores getDefaultActivateParameterDict
(unlike ActiveObject.activate).
Also, do not step on activate_kw, for readability.