ERP5Type/patches: Sort before trying to diff tuples in DiffTool
This is essential to have better aesthetic of Deepdiff diff for tuples.
In general, we treat iterables as long string so as to bypass the complicated
diff for iterables provided by DeepDiff, but for tuples this was creating a
problem as it was not possible to understand the diff properly.

Hence, by soting them, we get more understandable diff for tuples.

Also, there is no need to override `__diff_tuple` function as it is itself
dependent on `__diff_iterable` which we already override.