[erp5_officejs] Revive codemirror app for javascript only
Codemirror app is now using the same mechanism as text_editor. For now, it has
exactly the same features as text editor but it will create "Web Script"
instead of "Web Page" documents. The codemirror mode is forced to "javascript"
and the content type of the document if forced to "application/javascript".