DiffTool: Fixup default date time display in case we run diff action on an object
View wise, the date and path for new version/obvject should be on right side which
in our case is `your_second_date`. Thus, the case where we run diff action on single
object, the right side should show the current version and the left side shoule show
the last version.