testnode: setup SlapOS for test suite only at test startup
Our test nodes were installing seleniumrunner software in a loop, which
uses a lot of resources on the git repository and on the testnode
itself. Now that it's done only once at startup, this should reduce the
load a lot, but updates to seleniumrunnner profile needs to restart test
nodes.

Also add methods on taskdistributor.DummyTaskDistributor so that
ScalabilityTest slapos test node can be prepared.