[erp5_officejs] "Javascript Editor" is now a progressive webapp
Add manifests to make the app a progressive webapp. There are 2 manifests : one during the installation process (while the app is loading) and one for the app itself.
Note that the "start_url" parameter is made for the officejs app URL ([your_app].app.officejs.com/[release_number]). It doesn't work on an ERP5 instance ([instance_url]/erp5/web_site_module/[your_app]). If you want to make it work on the ERP5 instance for testing purposes, remove one "../".