[erp5_officejs_ui_test] Fix some OfficeJS UI tests
After clicking on "Documents", we need to wait for the listbox to be loaded before doing anything else. In particular, if we do a search after clicking on "Documents" the listbox is loaded twice (once after clicking on Documents and once after clicking on search). If we don't wait for the listbox to be loaded before doing the search, the search macro might end too early.