officejs_support_request_ui_test: snapshot test graphs
test that rendering of graphs match the reference image with new
assertCanvasImage selenium extension.

To get the reference png, we can right click on the canvas and "Save
Image as ...". It should not be too hard to extend assertCanvasImage
to be able to run with an "uplate reference snapshots" mode, but it's
not done at the moment.