Formulator: Do not add empty item if item_list is empty for SingleItemsWidget and MultiItemsWidget
we used to add this value because <select> without <option>
(that happens in MultiListField only) is invalid on XHTML.

but it was not at required for other cases like MultiCheckboxField or RadioField.
Also, <select> without <option> is no longer invalid on HTML5.

Considering that showing dummy unselectable <option> is anyway meaningless for users,
we drop it for all cases.