officejs_support_request_ui: review document ingestion
Publishing was too open. Setting metadata from the support request
project and the support request organisation seems a good default

Also ingest document in two steps:

* first step is to create the document and activate DMS metadata discovery
* second step is to share the document, once the necessary roles have been
  given to the user

What's still not good with this approach is that DMS ingestion might
merge this new document with an already existing document. This case
is not supported and needs an extension to ingestion API