crm: interaction to set SR stop date to now when closing
So that we can track more accurately the duration of the ticket.

The date is always set, erasing a date if it was set, for scenarios
where we close, reopen and close again. For cases where we forgot to
close and close later, we can always close and the change the date.