officejs_support_request_ui: Take into account preferred SR use
Only services whose use is included in the preferred SR use are
allowed when openning a support request.

We still filter the services according to supply lines applicable for
the selected project, but allow all SR services when project does not
define supplies.

Another user visible change is that test we no longer select a random
service, it must be a user decision (eventhough one possible
sophistication could be to select the service when there's only one).
Update test accordingly.

Also use relative URL for service, not just the service ID (this is API
breaking change, but the javascript part should not have this cached)