testTradeReports: Allow reusing in site-dependent security group setups.
This test alone does not define site-dependent security groups, so this
change is not required as such.
But in real configuration, ERP5 instance define security groups, which
sometimes (often ?) depend on which site a document or user is linked to.
When multiple groups apply (ex: having a certain function at a certain
site), the final group is composed by joining group identifiers using
underscores. This forbids using underscores in group identifiers.
And group identifiers are generated from document's (here, the Category)
codification if set, otherwise the reference if set, otherwise the document
identifier.
So set a underscore-less codification on these categories so this test can
be run on tests where security is configured to include site in security
groups.