officejs_support_request_ui: reuse existing API script for "Comment {Date,Author}"
Because this API script returns JSON, it cannot easily be used
"internally" server side, so the logic as been moved to a
SupportRequest_getCommentPostList and
SupportRequest_getCommentPostListAsJson just does the serialization to
JSON part.

It has been found that there is no convention for "scripts returning
JSON but that can also be called by other script", so this way is not an
example to follow.