accounting: restrict restarting accounting period to Assignor
This workflow involved an Assignee who can open, close, re-open and an
Assignor who can close definitively.
This is usually configured so that accountants are Assignee and CFO is
Assignor.

We realized that re-opening a Period that was previously closed is
something we don't want the accountants to do without CFO's approval.

To support this configuration, we only allow Assignor to re-open.

Now Assignee can open and close temporarily and Assignor can re-open and
close definitively.