Revert "Show Id of template."
This reverts commit b7d9590c9262f58d30178dddc278ba77b15b1a4d.

Users are not supposed to edit templates because documents are not in
their "normal" context, if user edit templates some unexpected things
might happen. We don't want user to change state or do some actions that
may cause interractions on other documents from a template. This was
never strongly enforced, but there was not supposed to be an easy way
for users to access the template document.
With old versions of ERP5 (before the rework of isIndexable /
isSubTreeIndexable ), users could also add lines inside the template
document and these lines where indexable.

Note that the way of editing templates is:
 * make a "real" document from the template
 * modify this document
 * use "Make Template" again - this will replace the document with the
same title.