discussion: remove temporary hidden input added for click
When clicking "Reply" on a discussion post, this script adds a hidden input to mark which post we are replying to.
This cause problem  when ctrl+clicking reply to reply in a new tab, because it leaves the page with the extra hidden input, which cause problems, like clicking reply again cause Error Value: You are not allowed to access '['1', '1']' in this context , or clicking change page will reply.

This is a minimal effort fix, just remove the added element after click.