[erp5_web_renderjs_ui] Display module actions in panel
Hardcodes a panel "VIEW: View" to hide "consistency", "history" and "metadata".

Also hides "ACTIONS: Change State" ("mass_workflow_jio") as it does not work yet on the panel,
this workaround can be removed once mass actions are moved to listbox actions instead of object_jio_action.

This commit makes a translation test to fail (testPanel).
The test was not waiting enough time before checking a value
in the "Workflow" section from the menu panel. It means an
additional async call is made somehow which should not be the
case for this testPanel.

/reviewed-on https://lab.nexedi.com/nexedi/erp5/merge_requests/673