erp5_web_renderjs_ui: Replace appcache by a service worker
Chrome will deprecate appcache in the coming weeks.
As appcache has been dropped on Firefox, this change will improve the speed on Firefox.

The service worker only provide a cache for all ERP5JS gadgets.
ERP5JS will still be usable on browser without service worker API.

The list of files to cache is directly added inside the service worker
code, in order to be able to skip the usage of the fetch method (which
does not yet cover all browser HTTP usage).

This list is configurable on the web site level.
No need to change the service worker code directly.

The service worker will be updated as soon as the web site modification date changes.

End user does not need to reload its browser tab.
Only browsing ERP5JS will trigger the service worker update.

When a new update is found, ERP5JS will automatically reload itself if
there is no risk of losing user data input.

An empty appcache is kept, to allow migration of existing client to the
worker automatically.