[erp5_hal_json_style/erp5_web_renderjs_ui] Add copy/paste actions
Activate copy/paste/delete buttons on form list and form view.

When selection listbox lines, only use what is visible / checked.
This is required because of list methods which do not support catalog queries.

This changes the behaviour of mass workflow transition action and delete action,
but at least, it is consistent everywhere now, and so, easier to explain (what you see is what you change).