task_distribution: better distribute the work depending on the priority of test suites
Before we were looking only create date of test results to give priority of work to do for a test nodes.
So we had cases where some test suites with medium priority had 5 testnodes, and test suites with
high priority having only 3 testnodes.

So clearly distribute the work depending on the maximum quantity of testnode we want.