ZSQLCatalog.SQLCatalog: Stop updating _max_uid .
Even if it does not lead to ConflictError being raised, updating _max_uid
has a cost and a transaction-serialising effect (most noticeable with NEO).
This value is not used (outside of a bootstrap value for portal_ids
generator), so stop doing this unnecessary work.
Also, fold produceUid into its only caller.
Also, change the folded pattern to react to an empty buffer rather than
test for emptiness before each access.