simulation: Test the divergence on Float Equivalence Tester with the epsilon-span by default
Before:
 Float Equivalence Tester does not test the divergence even if Divergence Provider is ON

After:
 Float Equivalence Tester test the divergence with the epsilon-span if Divergence Provider is ON

Before, we needed the following condition to test the divergence at
Float Equivalence Tester:

"Divergence Provider is ON"  AND
("Absolute Tolerance has some value"  OR
 "Relative Tolerance has some value" )

However this default behavior was not consistent with other testers, and
difficult to notice from the user interface.
Thus test the divergence by default with the epsilon span which is defined in
FloatEquivalenceTester class.

The epsilon-span is the same amount that when we put 0.0 into the ranges.
The 0.0 has been the configuration of erp5_configurator_standard_trade_template.