erp5_simulation: Reduce portal type duplication.
Solver Tool portal type is also present in erp5_base, and erp5_simulation
already depends on erp5_base.
The alternative would be to move this to erp5_simulation along with what
depends on it, which eventually includes workflows, which then break
erp5_accounting (which should not depend on erp5_simulation, despite
referencing workflows which do semantically belong to simulation).
This change is much simpler, sufficient to not cause a regression, and
this type is anyway already in erp5_base to begin with.