[erp5_web_renderjs_ui] Multirelationfield: fix editing
Only send line data if the text is not empty and ensure there is no hole
in the uid list.
This will fix the edition when the field already contains some homonyms.

Only append empty input if the latest one is not empty.

Replace verify by assert in the tests.