ZMySQLDA.db: Drop special handling of syntax errors.
The same error code can be used for different causes. Specifically, it can
be used for SQL-level syntax error (which this code intended to handle),
or it can be used for fulltext syntax error (which makes this code raise).
Also, drop forced-reconnection, because this means a still-established
connection make be kept around and completely detected from current
transaction, which may cause a database snapshot to be kept indefinitely
(actually, until mariadb's timeout), in turn causing performance issue
because of associated InnoDB snapshot.