ZMySQLDA.db: Close any replaced connection on reconnection.
MySQLdb relies on garbage collection to close the connection. Relying
on such asynchronous event as the garbage collector is never a good idea,
and this can cause issues when the transaction associated with the
connection is started: because the connector is being replaced, it will
never be properly aborted. So do not take chances and do not leak open
connections: always close it.