ZSQLCatalog: use _getProperty to skip security checks
During indexation we don't apply security checks, so this should be a
little bit faster.

With this change _getProperty becomes a more "official" API, so some
small changes had to be made to classes which do not inherits from
Products.ERP5Type.Base.Base, so that they also implements _getProperty:
 - for ERP5Site we simply use getProperty
 - for the test class from testERP5Catalog, the change is a bit more
important, because this class never defined getProperty, so during that
test we were just acquiering a getProperty from portal.