[erp5_hal_json_style/erp5_web_renderjs_ui] Render the form directly after a successfull call to Base_edit
This decreases the query number from 2 (POST to Base_edit -> redirect + GET to render the form) to 1 (POST to Base_edit -> render the form) and so improve the speed of ERP5JS.

To achieve this, apply so fixed in ERP5JS:
* rerender the panel/header after Base_edit (router is not called in such case, so, it must be manually called)
* fetch the page title from the form calculation (as the document context is not recalculated, only the form)
* force deleting the current field values, used by the relation field listbox search

This commit also force the full page refresh when diplaying a new document, in order to prevent keeping field values from the previous one (so bugs used to occur when using the previous/next document pagination).

Thanks to Tristan Cavelier for his work on this topic.