dms_ui_test: fix testContributeTab
 * Depending on conversion server speed, document will be Converted or
Converting. Assert that the state match Convert(ing|ed) to accept both.
 * prefer assert* instead of verify* to terminate test execution on
first failed assertion.
 * use more specific xpath instead of asserting that text is present in
the page.
 * checking that URL matches the previous URL interpreted as a regexp
was not working for IPv6 URL with brackets (because of the meaning of
regular expression). Because what we want here is to check that URL
starts with previously stored URL, check with a glob: expression, that
should be also compatible with IPv6 URLs.