erp5_pdm: It is never the good solution to delete matrix cells directly.
This breaks matrix index consistency.
Also, call isBasePricePerSlice rather than introspecting wrapped method
arguments - the only wrapped method is _setBasePricePerSlice anyway !
Also, trigger SupplyCell_updateSliceBasePrice as it depends on
isBasePricePerSlice value on this line, and it just changed, and that
workflow does not trigger on line changes (which is a bug).
Also, get rid of a trivial local variable.