accounting: Always show "Group Analytics" on Trial Balance Dialog
This field was only displayed if any of these was enabled:
 - by function analytics
 - by project analytics
 - by category analytics

and it allows requesting reports per function, project, categories and
section.

The problem is that if any analytics was explicitly enabled, we never
allow users to get a break down by section.

Always display the field, so that user can always select section.

This does not make much sense when there's only one section though.