[erp5_officejs] Fix error when converting new document with remote storage
In only office apps and notebook app download convert gadget produces error
when trying to convert a new document when using erp5 storage because it
redirects to the document with old jio_key which does not exist anymore
after syncing. The bug is fixed by introducing a cloudooo_only parameter
in ojs_sync gadget which is used to only sync cloudooo storage.