testnode: fix Computer.updateConfiguration call
updateConfiguration is a method to set the list of partitions, not to
add partitions to an existing computer. The previous (slapos.core
<1.5.0) implementation of slap proxy was incorrect and was adding more
partitions each time it was called, but that was never the indented
behavior.

Prepare the list of partitions and call updateConfiguration only once to
set all partitions.

This should fix SLAPOS-SR-TEST that are still using erp5.util