test_result: Support repositories or user ids with .
Because gitlab way of passing project ID as user/project URL-encoded
does not support dots (it would needs to be encoded as %2E) and that
requests rewrites %2F from URL to . (which is probably correct, that's
the RFC), we cannot use encoded paths for when they contain a dot.
To workaround this, when user or repo as a dot, make another API call to
get the numerical project ID and use that instead.
We only do that for projects with a dot because that's an extra request
we'd like to avoid.